refactor: merge thumbnail url data sources
