+++ src/data/source/implementation/UserAvatarThumbnailUrlDataSource.php
@@ -20,69 +20,16 @@
 
 namespace MediaWiki\Extension\RobloxAPI\data\source\implementation;
 
-use MediaWiki\Extension\RobloxAPI\data\args\ArgumentSpecification;
 use MediaWiki\Extension\RobloxAPI\data\source\DataSourceProvider;
-use MediaWiki\Extension\RobloxAPI\data\source\DependentDataSource;
-use MediaWiki\Extension\RobloxAPI\util\RobloxAPIUtil;
-use Parser;
+use MediaWiki\Extension\RobloxAPI\data\source\ThumbnailUrlDataSource;
 
-class UserAvatarThumbnailUrlDataSource extends DependentDataSource {
+class UserAvatarThumbnailUrlDataSource extends ThumbnailUrlDataSource {
 
 	/**
 	 * @inheritDoc
 	 */
 	public function __construct( DataSourceProvider $dataSourceProvider ) {
 		parent::__construct( $dataSourceProvider, 'userAvatarThumbnailUrl', 'userAvatarThumbnail' );
-	}
-
-	/**
-	 * @inheritDoc
-	 */
-	public function exec(
-		DataSourceProvider $dataSourceProvider, Parser $parser, array $requiredArgs, array $optionalArgs = []
-	) {
-		$data = $this->dataSource->exec( $dataSourceProvider, $parser, $requiredArgs );
-
-		if ( !$data ) {
-			return $this->failNoData();
-		}
-
-		if ( count( $data ) == 0 ) {
-			return $this->failInvalidData();
-		}
-
-		$url = $data[0]->imageUrl;
-
-		if ( !$url || !RobloxAPIUtil::verifyIsRobloxCdnUrl( $url ) ) {
-			return $this->failInvalidData();
-		}
-
-		$format = $optionalArgs['format'] ?? 'Png';
-		$lowerFormat = strtolower( $format );
-
-		return "$url.$lowerFormat";
-	}
-
-	/**
-	 * @inheritDoc
-	 */
-	public function shouldEscapeResult( $result ): bool {
-		// The url should not be escaped here in order to be embedded correctly using $wgEnableImageWhitelist.
-		// If the URL was escaped here, it would be URL-encoded and not recognized by MediaWiki as an image URL.
-		return !RobloxAPIUtil::verifyIsRobloxCdnUrl( $result );
-	}
-
-	/**
-	 * @inheritDoc
-	 */
-	public function getArgumentSpecification(): ArgumentSpecification {
-		return new ArgumentSpecification( [
-			'UserID',
-			'ThumbnailSize',
-		], [
-			'is_circular' => 'Boolean',
-			'format' => 'ThumbnailFormat',
-		], );
 	}
 
 	/**
